Fix permissions for show_details person columns
These fields should not be treated as public columns, one needs
show_details permission to view them.
In an event, people who can update the event may also view the
participants' full data. We do the same already for event questions, and
we have the add_requests mechanism in place to protect from accidentally
granting an event organizer access to a person's data.

Refs https://help.puzzle.ch/#ticket/zoom/5905